투표버튼갯수 일반화 작업 완료
옵션 4개일때 1번 2번 선택안되는 이슈있음
