+++ buy-or-not/Views/Main.swift
@@ -155,9 +155,7 @@ struct questionItem: View {
             
             ZStack {
                 RoundedRectangle(cornerRadius: 10).frame(height: 40).foregroundColor(Color.init(hex: "F2F2F7"))
-                HStack {
-                    voteButtons(options: options)
-                }
+                VoteButtonView(option:options)
             }.frame(height: 40)
             Spacer()
         }
@@ -166,13 +164,29 @@ struct questionItem: View {
     }
 }
 
-// 다중 버튼을 하기위해서 스테이트 변수를 컨트롤할 수 있는 다른 구조를 만들어야함
-struct voteButtons: View {
+
+struct VoteButtonView: View {
     
-    @State var options: [Options]
-    
-    @State var voteDone: Bool = false
+    @State var option: [Options]
+    // 옵션 최대 갯수 4개
     @State var buttonState: [Bool] = [false, false, false, false]
+    @State var voteDone: Bool = false
+
+
+    func buttonTab(index: Int) {
+        if buttonState[index] == true  {
+            self.voteDone = true
+        } else {
+            // 토글
+            for i in 0..<self.buttonState.count {
+                if (i == index) {
+                    self.buttonState[i] = true
+                } else {
+                    self.buttonState[i] = false
+                }
+            }
+        }
+    }
     
     var body: some View {
         if voteDone {
@@ -183,125 +197,23 @@ struct voteButtons: View {
                 // 투표 현황 텍스트 추가되어야함
             }
         } else {
-            switch options.count {
-            case 1:
-                Button(action: {
-                    if buttonState[0] {withAnimation{voteDone = true}} else {buttonState[0].toggle()}
-                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[0] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[0] ? "한번 더 누르시면\n투표가 반영됩니다." : "1번").foregroundColor(.black)
-                            .font(buttonState[0] ? .caption : .body)
-                    }
-                }
-            case 2:
-                HStack {
-                    Button(action: {
-                        if buttonState[0] {withAnimation{voteDone = true}} else {buttonState[0].toggle()}
-                        if buttonState[1] {buttonState[1].toggle()}
-                    }) {
+            HStack {
+                ForEach(0..<option.count) { index in
+                    Button {
+                        withAnimation {
+                            buttonTab(index: index)
+                        }
+                    } label: {
                         ZStack {
-                            Rectangle().foregroundColor(buttonState[0] ? .blue : .clear).cornerRadius(10)
-                            Text(buttonState[0] ? "한번 더 누르시면\n투표가 반영됩니다." : "1번").foregroundColor(.black)
-                                .font(buttonState[0] ? .caption : .body)
+                            Rectangle().foregroundColor(buttonState[index] ? .blue : .clear).cornerRadius(10)
+                            Text(buttonState[index] ? "한번 더 누르시면\n투표가 반영됩니다." : "\(index + 1)번").foregroundColor(.black)
+                                .font(buttonState[index] ? .caption : .body)
                         }
                     }
-                    Button(action: {
-                        if buttonState[1] {withAnimation{voteDone = true}} else {buttonState[1].toggle()}
-                        if buttonState[0] {buttonState[0].toggle()}
-                    }) {
-                        ZStack {
-                            Rectangle().foregroundColor(buttonState[1] ? .blue : .clear).cornerRadius(10)
-                            Text(buttonState[1] ? "한번 더 누르시면\n투표가 반영됩니다." : "2번").foregroundColor(.black)
-                                .font(buttonState[1] ? .caption : .body)
-                        }
-                    }
+
+
                 }
                 
-            case 3:
-                Button(action: {
-                    if buttonState[0] {withAnimation{voteDone = true}} else {buttonState[0].toggle()}
-                    if buttonState[1] {buttonState[1].toggle()}
-                    if buttonState[2] {buttonState[2].toggle()}
-                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[0] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[0] ? "한번 더 누르시면\n투표가 반영됩니다." : "1번").foregroundColor(.black)
-                            .font(buttonState[0] ? .caption : .body)
-                    }
-                }
-                Button(action: {
-                    if buttonState[1] {withAnimation{voteDone = true}} else {buttonState[1].toggle()}
-                    if buttonState[0] {buttonState[0].toggle()}
-                    if buttonState[2] {buttonState[2].toggle()}                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[1] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[1] ? "한번 더 누르시면\n투표가 반영됩니다." : "2번").foregroundColor(.black)
-                            .font(buttonState[1] ? .caption : .body)
-                    }
-                }
-                Button(action: {
-                    if buttonState[2] {withAnimation{voteDone = true}} else {buttonState[2].toggle()}
-                    if buttonState[1] {buttonState[1].toggle()}
-                    if buttonState[0] {buttonState[0].toggle()}
-                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[2] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[2] ? "한번 더 누르시면\n투표가 반영됩니다." : "3번").foregroundColor(.black)
-                            .font(buttonState[2] ? .caption : .body)
-                    }
-                }
-            case 4:
-                Button(action: {
-                    if buttonState[0] {withAnimation{voteDone = true}} else {buttonState[0].toggle()}
-                    if buttonState[1] {buttonState[1].toggle()}
-                    if buttonState[2] {buttonState[2].toggle()}
-                    if buttonState[3] {buttonState[3].toggle()}
-                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[0] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[0] ? "한번 더 누르시면\n투표가 반영됩니다." : "1번").foregroundColor(.black)
-                            .font(buttonState[0] ? .caption : .body)
-                    }
-                }
-                Button(action: {
-                    if buttonState[1] {withAnimation{voteDone = true}} else {buttonState[1].toggle()}
-                    if buttonState[0] {buttonState[0].toggle()}
-                    if buttonState[2] {buttonState[2].toggle()}
-                    if buttonState[3] {buttonState[3].toggle()}
-                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[1] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[1] ? "한번 더 누르시면\n투표가 반영됩니다." : "2번").foregroundColor(.black)
-                            .font(buttonState[1] ? .caption : .body)
-                    }
-                }
-                Button(action: {
-                    if buttonState[2] {withAnimation{voteDone = true}} else {buttonState[2].toggle()}
-                    if buttonState[1] {buttonState[1].toggle()}
-                    if buttonState[0] {buttonState[0].toggle()}
-                    if buttonState[3] {buttonState[3].toggle()}
-                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[2] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[2] ? "한번 더 누르시면\n투표가 반영됩니다." : "3번").foregroundColor(.black)
-                            .font(buttonState[2] ? .caption : .body)
-                    }
-                }
-                Button(action: {
-                    if buttonState[3] {withAnimation{voteDone = true}} else {buttonState[3].toggle()}
-                    if buttonState[1] {buttonState[1].toggle()}
-                    if buttonState[2] {buttonState[2].toggle()}
-                    if buttonState[0] {buttonState[0].toggle()}
-                }) {
-                    ZStack {
-                        Rectangle().foregroundColor(buttonState[3] ? .blue : .clear).cornerRadius(10)
-                        Text(buttonState[3] ? "한번 더 누르시면\n투표가 반영됩니다." : "4번").foregroundColor(.black)
-                            .font(buttonState[3] ? .caption : .body)
-                    }
-                }
-            default:
-                Text("버튼 최대갯수 초과")
             }
         }
     }
